requirements.txt
Pillow>=9.0.0
piexif>=1.1.3